+++ lib/coloraide/color/distance/delta_e_76.py
@@ -1,16 +1,26 @@
 """Delta E 76."""
-
+from ..distance import DeltaE
 from . import distance_euclidean
 
 
-def distance(color, sample, **kwargs):
-    """
-    Delta E 1976 color distance formula.
+class DE76(DeltaE):
+    """Delta E 76 class."""
 
-    http://www.brucelindbloom.com/index.html?Eqn_DeltaE_CIE76.html
+    @staticmethod
+    def name():
+        """Name of method."""
 
-    Basically this is Euclidean distance in the Lab space.
-    """
+        return "76"
 
-    # Equation (1)
-    return distance_euclidean.distance(color, sample, space="lab")
+    @staticmethod
+    def distance(color, sample, **kwargs):
+        """
+        Delta E 1976 color distance formula.
+
+        http://www.brucelindbloom.com/index.html?Eqn_DeltaE_CIE76.html
+
+        Basically this is Euclidean distance in the Lab space.
+        """
+
+        # Equation (1)
+        return distance_euclidean.distance(color, sample, space="lab")