+++ lib/coloraide/spaces/jzazbz.py
@@ -106,6 +106,11 @@ class Jzazbz(Labish, Space):
     SPACE = "jzazbz"
     SERIALIZE = ("--jzazbz",)
     CHANNEL_NAMES = ("jz", "az", "bz", "alpha")
+    CHANNEL_ALIASES = {
+        "lightness": 'jz',
+        "a": 'az',
+        "b": 'bz'
+    }
     DEFAULT_MATCH = re.compile(RE_DEFAULT_MATCH.format(color_space='|'.join(SERIALIZE), channels=3))
     WHITE = "D65"